+++ src/components/Navbar.tsx
@@ -303,15 +303,15 @@ const Navbar = () => {
                       </div>
                       <div className="py-1">
                         <Link
-                          to="/signin"
+                          to="/auth?tab=signin"
                           className="block px-4 py-2 text-sm text-foreground hover:bg-accent hover:text-accent-foreground"
                           onClick={() => setIsUserMenuOpen(false)}
                         >
                           Sign in
                         </Link>
                         <Link
-                          to="/signup"
-                          state={{ tab: 'signup' }}
+                          to="/auth?tab=signup"
+                        
                           className="block px-4 py-2 text-sm text-foreground hover:bg-accent hover:text-accent-foreground"
                           onClick={() => setIsUserMenuOpen(false)}
                         >
