+++ src/components/ui/tabs.tsx
@@ -11,6 +11,7 @@ const TabsList = React.forwardRef<
   <TabsPrimitive.List
     ref={ref}
     className={cn(
+      // Default tab list background — light muted tone
       'inline-flex h-10 items-center justify-center rounded-md bg-muted p-1 text-muted-foreground',
       className
     )}
@@ -27,8 +28,8 @@ const TabsTrigger = React.forwardRef<
     ref={ref}
     className={cn(
       'inline-flex items-center justify-center whitespace-nowrap rounded-sm px-3 py-1.5 text-sm font-medium ring-offset-background transition-all focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-ring focus-visible:ring-offset-2 disabled:pointer-events-none disabled:opacity-50 ' +
-        // 🔥 make active tab more visible
-        'data-[state=active]:bg-primary data-[state=active]:text-primary-foreground data-[state=active]:font-semibold data-[state=active]:shadow-md data-[state=active]:ring-1 data-[state=active]:ring-primary/70',
+        // 🎨 Subtle active state — only slightly darker than bg-muted
+        'data-[state=active]:bg-gray-200 data-[state=active]:text-foreground data-[state=active]:font-semibold data-[state=active]:shadow-sm',
       className
     )}
     {...props}
