+++ src/pages/Auth.tsx
@@ -1,0 +1,307 @@
+import React, { useState } from 'react';
+import { useNavigate,useSearchParams } from 'react-router-dom';
+import { useAuth } from '@/contexts/AuthContext';
+import { Button } from '@/components/ui/button';
+import { Input } from '@/components/ui/input';
+import { Label } from '@/components/ui/label';
+import {
+  Card,
+  CardContent,
+  CardFooter,
+  CardHeader,
+  CardTitle,
+  CardDescription,
+} from '@/components/ui/card';
+import { Tabs, TabsList, TabsTrigger, TabsContent } from '@/components/ui/tabs';
+import { Eye, EyeOff } from 'lucide-react';
+import { toast } from 'sonner';
+import Navbar from '@/components/Navbar';
+import Footer from '@/components/Footer';
+import { useEffect } from 'react';
+
+const AuthPage = () => {
+  const { user, signIn, signUp, isLoading } = useAuth();
+  const navigate = useNavigate();
+
+  const [searchParams] = useSearchParams();
+  const [tab, setTab] = useState<'signin' | 'signup'>(
+    (searchParams.get('tab') as 'signin' | 'signup') || 'signin'
+  );
+
+  useEffect(() => {
+    const tabFromUrl = (searchParams.get('tab') as 'signin' | 'signup') || 'signin';
+    setTab(tabFromUrl);
+  }, [searchParams]);
+
+
+  // Common states
+  const [email, setEmail] = useState('');
+  const [password, setPassword] = useState('');
+  const [showPassword, setShowPassword] = useState(false);
+  const [isSubmitting, setIsSubmitting] = useState(false);
+
+  // Signup-specific states
+  const [firstName, setFirstName] = useState('');
+  const [lastName, setLastName] = useState('');
+  const [confirmPassword, setConfirmPassword] = useState('');
+  const [showConfirm, setShowConfirm] = useState(false);
+
+  // -----------------------
+  // LOGIN HANDLER
+  // -----------------------
+  const handleLogin = async (e: React.FormEvent) => {
+    e.preventDefault();
+    if (!email || !password) {
+      toast.error('Please enter email and password');
+      return;
+    }
+
+    setIsSubmitting(true);
+    const { error } = await signIn(email, password);
+    setIsSubmitting(false);
+
+    if (error) toast.error(error.message || 'Login failed');
+    else toast.success('Login successful!');
+  };
+
+  // -----------------------
+  // SIGNUP HANDLER
+  // -----------------------
+  const handleSignup = async (e: React.FormEvent) => {
+    e.preventDefault();
+    if (!email || !password) {
+      toast.error('Please enter email and password');
+      return;
+    }
+
+    if (password !== confirmPassword) {
+      toast.error('Passwords do not match');
+      return;
+    }
+
+    if (password.length < 6) {
+      toast.error('Password must be at least 6 characters');
+      return;
+    }
+
+    setIsSubmitting(true);
+    const userData = { first_name: firstName, last_name: lastName };
+    const { error } = await signUp(email, password, userData);
+    setIsSubmitting(false);
+
+    if (error) toast.error(error.message || 'Signup failed');
+    else {
+      toast.success('Account created! Please check your email for verification.');
+      setFirstName('');
+      setLastName('');
+      setEmail('');
+      setPassword('');
+      setConfirmPassword('');
+    }
+  };
+
+  return (
+    <div className="min-h-screen flex flex-col bg-gradient-to-b from-background/80 to-background">
+      <Navbar />
+      <div className="flex-grow flex items-center justify-center p-4">
+        <div className="w-full max-w-md">
+          <Card className="border shadow-lg">
+            <CardHeader className="space-y-1 text-center">
+              <CardTitle className="text-2xl font-bold">
+                {tab === 'signin' ? 'Welcome Back' : 'Create Account'}
+              </CardTitle>
+              <CardDescription >
+                {tab === 'signin'
+                  ? 'Sign in to your account'
+                  : 'Sign up to get started'}
+              </CardDescription>
+
+              {/* Tabs Navigation */}
+             <Tabs
+  value={tab}
+  onValueChange={(val: any) => {
+    setTab(val);
+    navigate(`/auth?tab=${val}`); // keeps URL synced when switching tabs
+  }}
+  className="w-full"
+>
+                <TabsList className="grid grid-cols-2 w-full mt-2">
+                  <TabsTrigger value="signin">Signin</TabsTrigger>
+                  <TabsTrigger value="signup">Signup</TabsTrigger>
+                </TabsList>
+
+                {/* ----------------------- */}
+                {/* SIGN IN TAB CONTENT */}
+                {/* ----------------------- */}
+                <TabsContent value="signin">
+                  <form onSubmit={handleLogin} className='text-left'>
+                    <CardContent className="space-y-4 mt-4">
+                      <div className="space-y-2">
+                        <Label htmlFor="email">Email</Label>
+                        <Input
+                          id="email"
+                          type="email"
+                          placeholder="name@example.com"
+                          value={email}
+                          onChange={(e) => setEmail(e.target.value)}
+                          required
+                        />
+                      </div>
+
+                      <div className="space-y-2">
+                        <Label htmlFor="password">Password</Label>
+                        <div className="relative">
+                          <Input
+                            id="password"
+                            type={showPassword ? 'text' : 'password'}
+                            placeholder="••••••••"
+                            value={password}
+                            onChange={(e) => setPassword(e.target.value)}
+                            required
+                            className="pr-10"
+                          />
+                          <button
+                            type="button"
+                            onClick={() => setShowPassword(!showPassword)}
+                            className="absolute right-3 top-1/2 -translate-y-1/2 text-gray-500 hover:text-gray-700 focus:outline-none"
+                          >
+                            {showPassword ? (
+                              <EyeOff className="h-4 w-4" />
+                            ) : (
+                              <Eye className="h-4 w-4" />
+                            )}
+                          </button>
+                        </div>
+                      </div>
+                    </CardContent>
+
+                    <CardFooter>
+                      <Button
+                        type="submit"
+                        className="w-full bg-primary"
+                        disabled={isSubmitting || isLoading}
+                      >
+                        {isSubmitting ? 'Signing in...' : 'Sign In'}
+                      </Button>
+                    </CardFooter>
+                  </form>
+                </TabsContent>
+
+                {/* ----------------------- */}
+                {/* SIGN UP TAB CONTENT */}
+                {/* ----------------------- */}
+                <TabsContent value="signup">
+                  <form onSubmit={handleSignup} className='text-left'>
+                    <CardContent className="space-y-4 mt-4">
+                      <div className="grid grid-cols-2 gap-4">
+                        <div className="space-y-2">
+                          <Label htmlFor="first-name">First name</Label>
+                          <Input
+                            id="first-name"
+                            placeholder="John"
+                            value={firstName}
+                            onChange={(e) => setFirstName(e.target.value)}
+                            required
+                          />
+                        </div>
+                        <div className="space-y-2">
+                          <Label htmlFor="last-name">Last name</Label>
+                          <Input
+                            id="last-name"
+                            placeholder="Doe"
+                            value={lastName}
+                            onChange={(e) => setLastName(e.target.value)}
+                            required
+                          />
+                        </div>
+                      </div>
+
+                      <div className="space-y-2">
+                        <Label htmlFor="email">Email</Label>
+                        <Input
+                          id="email"
+                          type="email"
+                          placeholder="name@example.com"
+                          value={email}
+                          onChange={(e) => setEmail(e.target.value)}
+                          required
+                        />
+                      </div>
+
+                      <div className="space-y-2">
+                        <Label htmlFor="password">Password</Label>
+                        <div className="relative">
+                          <Input
+                            id="password"
+                            type={showPassword ? 'text' : 'password'}
+                            placeholder="••••••••"
+                            value={password}
+                            onChange={(e) => setPassword(e.target.value)}
+                            required
+                            className="pr-10"
+                          />
+                          <button
+                            type="button"
+                            onClick={() => setShowPassword(!showPassword)}
+                            className="absolute right-3 top-1/2 -translate-y-1/2 text-gray-500 hover:text-gray-700 focus:outline-none"
+                            aria-label={showPassword ? 'Hide password' : 'Show password'}
+                          >
+                            {showPassword ? (
+                              <EyeOff className="h-4 w-4" />
+                            ) : (
+                              <Eye className="h-4 w-4" />
+                            )}
+                          </button>
+                        </div>
+                      </div>
+
+                      <div className="space-y-2">
+                        <Label htmlFor="confirm-password">Confirm Password</Label>
+                        <div className="relative">
+                          <Input
+                            id="confirm-password"
+                            type={showConfirm ? 'text' : 'password'}
+                            placeholder="••••••••"
+                            value={confirmPassword}
+                            onChange={(e) => setConfirmPassword(e.target.value)}
+                            required
+                            className="pr-10"
+                          />
+                          <button
+                            type="button"
+                            onClick={() => setShowConfirm(!showConfirm)}
+                            className="absolute right-3 top-1/2 -translate-y-1/2 text-gray-500 hover:text-gray-700 focus:outline-none"
+                            aria-label={showConfirm ? 'Hide password' : 'Show password'}
+                          >
+                            {showConfirm ? (
+                              <EyeOff className="h-4 w-4" />
+                            ) : (
+                              <Eye className="h-4 w-4" />
+                            )}
+                          </button>
+                        </div>
+                      </div>
+                    </CardContent>
+
+                    <CardFooter>
+                      <Button
+                        type="submit"
+                        className="w-full bg-primary"
+                        disabled={isSubmitting || isLoading}
+                      >
+                        {isSubmitting ? 'Creating Account...' : 'Create Account'}
+                      </Button>
+                    </CardFooter>
+                  </form>
+                </TabsContent>
+              </Tabs>
+            </CardHeader>
+          </Card>
+        </div>
+      </div>
+      <Footer />
+    </div>
+  );
+};
+
+export default AuthPage;
