--- src/pages/Login.tsx
@@ -1,127 +1,0 @@
-import React, { useState } from 'react';
-import { Navigate, useLocation, useNavigate } from 'react-router-dom';
-import { useAuth } from '@/contexts/AuthContext';
-import { Button } from '@/components/ui/button';
-import { Input } from '@/components/ui/input';
-import { Label } from '@/components/ui/label';
-import { Card, CardContent, CardFooter, CardHeader, CardTitle, CardDescription } from '@/components/ui/card';
-import { Tabs, TabsList, TabsTrigger } from '@/components/ui/tabs';
-import { Eye, EyeOff } from 'lucide-react';
-import { toast } from 'sonner';
-import Navbar from '@/components/Navbar';
-import Footer from '@/components/Footer';
-
-const LoginPage = () => {
-  const { user, signIn, isLoading } = useAuth();
-  const location = useLocation();
-  const navigate = useNavigate();
-
-  const [email, setEmail] = useState('');
-  const [password, setPassword] = useState('');
-  const [showPassword, setShowPassword] = useState(false);
-  const [isSubmitting, setIsSubmitting] = useState(false);
-
-  const from = location.state?.from?.pathname || '/';
-
-  if (user && !isLoading) {
-    return <Navigate to={from} replace />;
-  }
-
-  const handleLogin = async (e: React.FormEvent) => {
-    e.preventDefault();
-    if (!email || !password) {
-      toast.error('Please enter email and password');
-      return;
-    }
-
-    setIsSubmitting(true);
-    const { error } = await signIn(email, password);
-    setIsSubmitting(false);
-
-    if (error) toast.error(error.message || 'Login failed');
-  };
-
-  return (
-    <div className="min-h-screen flex flex-col bg-gradient-to-b from-background/80 to-background">
-      <Navbar />
-      <div className="flex-grow flex items-center justify-center p-4">
-        <div className="w-full max-w-md">
-          <Card className="border shadow-lg">
-            <CardHeader className="space-y-1 text-center">
-              <CardTitle className="text-2xl font-bold">Welcome Back</CardTitle>
-              <CardDescription>Sign in to your account</CardDescription>
-
-              {/* Tabs Navigation */}
-             <Tabs defaultValue="signin" className="w-full">
-
-                <TabsList className="grid grid-cols-2 w-full">
-                  <TabsTrigger
-                    value="signin"
-                    onClick={() => navigate('/signin')}
-                    className="cursor-pointer"
-                  >
-                    Signin
-                  </TabsTrigger>
-                  <TabsTrigger
-                    value="signup"
-                    onClick={() => navigate('/signup')}
-                    className="cursor-pointer"
-                  >
-                    Signup
-                  </TabsTrigger>
-                </TabsList>
-              </Tabs>
-            </CardHeader>
-
-            <form onSubmit={handleLogin}>
-              <CardContent className="space-y-4">
-                <div className="space-y-2">
-                  <Label htmlFor="email">Email</Label>
-                  <Input
-                    id="email"
-                    type="email"
-                    placeholder="name@example.com"
-                    value={email}
-                    onChange={e => setEmail(e.target.value)}
-                    required
-                  />
-                </div>
-
-                <div className="space-y-2">
-                  <Label htmlFor="password">Password</Label>
-                  <div className="relative">
-                    <Input
-                      id="password"
-                      type={showPassword ? 'text' : 'password'}
-                      placeholder="••••••••"
-                      value={password}
-                      onChange={e => setPassword(e.target.value)}
-                      required
-                      className="pr-10"
-                    />
-                    <button
-                      type="button"
-                      onClick={() => setShowPassword(!showPassword)}
-                      className="absolute right-3 top-1/2 -translate-y-1/2 text-gray-500 hover:text-gray-700 focus:outline-none"
-                    >
-                      {showPassword ? <EyeOff className="h-4 w-4" /> : <Eye className="h-4 w-4" />}
-                    </button>
-                  </div>
-                </div>
-              </CardContent>
-
-              <CardFooter>
-                <Button type="submit" className="w-full bg-primary" disabled={isSubmitting || isLoading}>
-                  {isSubmitting ? 'Signing in...' : 'Sign In'}
-                </Button>
-              </CardFooter>
-            </form>
-          </Card>
-        </div>
-      </div>
-      <Footer />
-    </div>
-  );
-};
-
-export default LoginPage;
